requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
oauthlib==3.1.1
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
